requirements.txt
awscli==1.11.2
beautifulsoup4==4.5.1
boto3==1.4.1
botocore==1.4.60
colorama==0.3.7
decorator==4.0.10
defusedxml==0.4.1
Django==1.10.2
django-allauth==0.27.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.6
django-filter==0.15.2
django-gravatar2==1.4.0
django-webpack-loader==0.3.3
djangorestframework==3.4.7
docutils==0.12
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
jmespath==0.9.0
oauthlib==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
pyasn1==0.1.9
Pygments==2.1.3
python-dateutil==2.5.3
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
rsa==3.4.2
s3transfer==0.1.7
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.1
traitlets==4.3.1
wcwidth==0.1.7
